Fixed MS VS compilation warning. JIRA CXX-7879 CXX-9262
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@77854 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/connect/services/netservice_params.hpp
@@ -239,7 +239,7 @@ protected:
 
 private:
     CCompoundRegistry m_Registry;
-    size_t m_Priority = 0;
+    int m_Priority = 0;
 };
 
 using CSynRegistry = TSynRegistry<CSynRegistryImpl>;
